test code fixes
