+++ MageTests/MAGETests-Bridging-Header.h
@@ -1,3 +1,4 @@
 #import <KIF/KIF.h>
+#import "MAGE-Bridging-Header.h"
 
 #import "TestingAppDelegate.h"
