+++ MageTests/Observation/Edit/ObservationEditCardCollectionViewControllerTests.swift
@@ -9,7 +9,6 @@
 import Foundation
 import Quick
 import Nimble
-//import Nimble_Snapshots
 
 @testable import MAGE
 import CoreData
@@ -20,7 +19,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
         
         describe("ObservationEditCardCollectionViewController") {
             var observationEditController: ObservationEditCardCollectionViewController!
-            var view: UIView!
             var window: UIWindow!;
             var stackSetup = false;
             
@@ -30,10 +28,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                     stackSetup = true;
                 }
                 window = TestHelpers.getKeyWindowVisible();
-                
-//                Nimble_Snapshots.setNimbleTolerance(0.0);
-//                Nimble_Snapshots.recordAllSnapshots()
-                
                 TestHelpers.resetUserDefaults();
 
                 MageCoreDataFixtures.clearAllData();
@@ -48,7 +42,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController.dismiss(animated: false);
                 window.rootViewController = nil;
                 observationEditController = nil;
-                view = nil;
             }
             
             describe("Legacy") {
@@ -68,9 +61,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                     observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: false, containerScheme: MAGEScheme.scheme());
                     
                     window.rootViewController = observationEditController;
-                    view = observationEditController.view;
-                    
-//                    expect(view).to(haveValidSnapshot(usesDrawRect: true));
                 }
                 
                 it("verify legacy behavior") {
@@ -84,7 +74,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                     
                     let nc = UINavigationController(rootViewController: observationEditController);
                     window.rootViewController = nc;
-                    view = observationEditController.view;
                     
                     tester().waitForView(withAccessibilityLabel: "attachments Gallery");
                     tester().waitForView(withAccessibilityLabel: "Edit Attachment Card")
@@ -135,10 +124,8 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 let nc = UINavigationController(rootViewController: observationEditController);
                 
                 window.rootViewController = nc;
-                view = observationEditController.view;
                 
                 expect(observationEditController.title) == "Edit Observation";
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
             it("empty new observation zero forms") {
@@ -151,8 +138,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
             it("validation error on observation") {
@@ -165,7 +150,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 let nc = UINavigationController(rootViewController: observationEditController);
                 window.rootViewController = nc;
-                view = observationEditController.view;
                 
                 TestHelpers.printAllAccessibilityLabelsInWindows();
                 tester().tapView(withAccessibilityLabel: "Save");
@@ -182,13 +166,11 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form");
                 
                 expect(delegate.addFormCalled).to(beTrue());
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
             it("show the form button if there are two forms") {
@@ -201,7 +183,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form");
@@ -220,7 +201,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
 //                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
@@ -234,7 +214,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form");
                 expect(delegate.addFormCalled).to(beTrue());
@@ -250,7 +229,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form")
@@ -281,7 +259,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form")
@@ -308,7 +285,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form")
@@ -342,7 +318,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 let nc = UINavigationController(rootViewController: observationEditController);
                 
                 window.rootViewController = nc;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form")
@@ -384,7 +359,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 let nc = UINavigationController(rootViewController: observationEditController);
                 
                 window.rootViewController = nc;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form")
@@ -437,7 +411,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 let nc = UINavigationController(rootViewController: observationEditController);
                 
                 window.rootViewController = nc;
-                view = observationEditController.view;
                 
                 // try to save with zero forms, should fail
                 tester().waitForTappableView(withAccessibilityLabel: "Save")
@@ -485,7 +458,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 let nc = UINavigationController(rootViewController: observationEditController);
                 
                 window.rootViewController = nc;
-                view = observationEditController.view;
                 
                 // try to save with zero forms, should fail
                 tester().waitForTappableView(withAccessibilityLabel: "Save")
@@ -548,8 +520,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: false, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
             it("observation should expand current forms") {
@@ -585,12 +555,10 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: false, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 
                 tester().waitForView(withAccessibilityLabel: "expand");
                 tester().tapView(withAccessibilityLabel: "expand");
                 tester().waitForAnimationsToFinish();
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
             it("observation should show current forms multiple forms") {
@@ -630,8 +598,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: false, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
             it("observation should show all the things form") {
@@ -667,8 +633,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
             it("observation should show checkbox form") {
@@ -682,7 +646,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form")
@@ -692,7 +655,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                     observationEditController.formAdded(form: (event.forms!)[0]);
                 }
                 
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
             it("filling out the form should update the form header") {
@@ -706,7 +668,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form")
@@ -720,7 +681,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 tester().tapView(withAccessibilityLabel: "Done");
                 tester().clearText(fromAndThenEnterText: "Some other text", intoViewWithAccessibilityLabel: "field1");
                 tester().tapView(withAccessibilityLabel: "Done");
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
             
             it("saving the form should send the observation to the delegate") {
@@ -735,7 +695,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 let navigationController = UINavigationController(rootViewController: observationEditController);
                 
                 window.rootViewController = navigationController;
-                view = observationEditController.view;
                 
                 tester().waitForView(withAccessibilityLabel: "ObservationEditCardCollection");
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
@@ -790,7 +749,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 let navigationController = UINavigationController(rootViewController: observationEditController);
                 
                 window.rootViewController = navigationController;
-                view = observationEditController.view;
                 
                 tester().waitForView(withAccessibilityLabel: "ObservationEditCardCollection");
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
@@ -833,7 +791,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 observationEditController = ObservationEditCardCollectionViewController(delegate: delegate, observation: observation, newObservation: true, containerScheme: MAGEScheme.scheme());
                 
                 window.rootViewController = observationEditController;
-                view = observationEditController.view;
                 
                 tester().waitForTappableView(withAccessibilityLabel: "Add Form");
                 tester().tapView(withAccessibilityLabel: "Add Form")
@@ -846,7 +803,6 @@ class ObservationEditCardCollectionViewControllerTests: KIFSpec {
                 tester().setText("", intoViewWithAccessibilityLabel: "field1");
                 (viewTester().usingLabel("Field View field0").view as? TextFieldView)?.textFieldDidEndEditing(viewTester().usingLabel("field0").view as! UITextField)
                 (viewTester().usingLabel("Field View field1").view as? TextFieldView)?.textFieldDidEndEditing(viewTester().usingLabel("field1").view as! UITextField)
-//                expect(view).to(haveValidSnapshot(usesDrawRect: true));
             }
         }
     }
