+++ MageTests/Observation/Edit/ObservationEditCoordinatorTests.swift
@@ -9,6 +9,7 @@
 import Foundation
 import Quick
 import Nimble
+import sf_ios
 
 import MagicalRecord
